completed modularization and creating a goose
