--- animals.py
@@ -1,170 +1,0 @@
-from datetime import date
-
-
-class Animal:
-    def __init__(self, name, species, food, chip_num):
-        self.name = name
-        self.species = species
-        self.food = food
-        self.__chip_number = chip_num
-        self.date_added = date.today()
-
-    def feed(self):
-        print(f'{self.name} was fed {self.food} on {date.today().strftime("%m/%d/%Y")}')
-
-    @property
-    def chip_number(self):
-        return self.__chip_number
-
-    @chip_number.setter
-    def chip_number(self, num):
-        pass
-
-
-#  WALKING ANIMALS
-
-class Llama(Animal):
-    def __init__(self, name, species, shift, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.shift = shift
-        self.walking = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-    def feed(self):
-        print(f'{self.name} was fed {self.food} while singing "Rockytop" on {date.today().strftime("%m/%d/%Y")}')
-
-
-class Giraffe(Animal):
-    def __init__(self, name, species, shift, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.shift = shift
-        self.walking = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Goat(Animal):
-    def __init__(self, name, species, shift, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.shift = shift
-        self.walking = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Burro(Animal):
-    def __init__(self, name, species, shift, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.shift = shift
-        self.walking = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Pig(Animal):
-    def __init__(self, name, species, shift, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.shift = shift
-        self.walking = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-#  SWIMMING ANIMALS
-
-class Otter(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.swimming = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Fish(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.swimming = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Duck(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.swimming = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Alligator(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.swimming = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Turtle(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.swimming = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-# Slithering Animals
-
-class Snake(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.slithering = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Snail(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.slithering = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Worm(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.slithering = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Slug(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.slithering = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
-
-
-class Lizard(Animal):
-    def __init__(self, name, species, food, chip_num):
-        super().__init__(name, species, food, chip_num)
-        self.slithering = True
-
-    def __str__(self):
-        return f"{self.name} is a {self.species}"
