+++ animals/__init__.py
@@ -1,0 +1,16 @@
+from .alligator import Alligator
+from .burro import Burro
+from .duck import Duck
+from .fish import Fish
+from .giraffe import Giraffe
+from .goat import Goat
+from .goose import Goose
+from .lizard import Lizard
+from .llama import Llama
+from .otter import Otter
+from .pig import Pig
+from .slug import Slug
+from .snake import Snake
+from .snail import Snail
+from .turtle import Turtle
+from .worm import Worm
