+++ animals/duck.py
@@ -1,0 +1,7 @@
+from .animal import Animal
+
+
+class Duck(Animal):
+    def __init__(self, name, species, food, chip_num):
+        super().__init__(name, species, food, chip_num)
+        self.walking = True
