+++ movements/__init__.py
@@ -1,0 +1,3 @@
+from .walking import Walking
+from .swimming import Swimming
+from .slithering import Slithering
