+++ movements/slithering.py
@@ -1,0 +1,8 @@
+class Slithering:
+
+    def __init__(self):
+        self.slither_speed = 0
+        self.length = 0
+
+    def slither(self):
+        print("The animal slithers")
