+++ movements/swimming.py
@@ -1,0 +1,8 @@
+class Swimming:
+
+    def __init__(self):
+        self.swim_speed = 0
+        self.maximum_depth = 0
+
+    def swim(self):
+        print(f"{self} swims")
