+++ movements/walking.py
@@ -1,0 +1,8 @@
+class Walking:
+
+    def __init__(self):
+        self.walk_speed = 0
+        self.legs = 0
+
+    def run(self):
+        print("The animal walks")
